+++ src/test/kotlin/Tests.kt
@@ -35,7 +35,7 @@ class Tests {
         val expected = listOf(
             "something:dev-1.1.0-alpha3(T2):1.2.14",
             "de.mpicbg.scicomp:kutils:0.7",
-            "com.github.holgerbrandl:kscript-annotations:1.2"
+            "com.github.holgerbrandl:kscript-annotations:1.4"
         )
 
         Script(lines).collectDependencies() shouldBe expected
@@ -46,7 +46,6 @@ class Tests {
         //            extractDependencies("""@file:DependsOn("com.squareup.moshi:moshi:1.5.0,com.squareup.moshi:moshi-adapters:1.5.0")""")
         //        }
     }
-
 
     @Test
     fun mixedDependencyCollect() {
@@ -59,7 +58,7 @@ class Tests {
             "de.mpicbg.scicomp.joblist:joblist-kotlin:1.1",
             "de.mpicbg.scicomp:kutils:0.7",
             "log4j:log4j:1.2.14",
-            "com.github.holgerbrandl:kscript-annotations:1.2"
+            "com.github.holgerbrandl:kscript-annotations:1.4"
         )
 
         Script(lines).collectDependencies() shouldBe expected
@@ -84,7 +83,42 @@ class Tests {
             collectDependencies() shouldBe listOf(
                 "net.clearvolume:cleargl:2.0.1",
                 "log4j:log4j:1.2.14",
-                "com.github.holgerbrandl:kscript-annotations:1.2"
+                "com.github.holgerbrandl:kscript-annotations:1.4"
+            )
+        }
+
+    }
+
+    @Test
+    fun customRepoWithCreds() {
+        val lines = listOf(
+                """@file:MavenRepository("imagej-releases", "http://maven.imagej.net/content/repositories/releases", user="user", password="pass") """,
+                // Same but name arg comes last
+                """@file:MavenRepository("imagej-snapshots", "http://maven.imagej.net/content/repositories/snapshots", password="pass", user="user") """,
+                // Whitespaces around credentials see #228
+                """@file:MavenRepository("spaceAroundCredentials", "http://maven.imagej.net/content/repositories/snapshots", password= "pass" , user= "user" ) """,
+                // Different whitespaces around credentials see #228
+                """@file:MavenRepository("spaceAroundCredentials2", "http://maven.imagej.net/content/repositories/snapshots", password= "pass", user="user" ) """,
+
+                // some other script bits unrelated to the repo definition
+                """@file:DependsOnMaven("net.clearvolume:cleargl:2.0.1")""",
+                """@file:DependsOn("log4j:log4j:1.2.14")""",
+                """println("foo")"""
+        )
+
+        with(Script(lines)) {
+
+            collectRepos() shouldBe listOf(
+                    MavenRepo("imagej-releases", "http://maven.imagej.net/content/repositories/releases", "user", "pass"),
+                    MavenRepo("imagej-snapshots", "http://maven.imagej.net/content/repositories/snapshots", "user", "pass"),
+                    MavenRepo("spaceAroundCredentials", "http://maven.imagej.net/content/repositories/snapshots", "user", "pass"),
+                    MavenRepo("spaceAroundCredentials2", "http://maven.imagej.net/content/repositories/snapshots", "user", "pass")
+            )
+
+            collectDependencies() shouldBe listOf(
+                    "net.clearvolume:cleargl:2.0.1",
+                    "log4j:log4j:1.2.14",
+                    "com.github.holgerbrandl:kscript-annotations:1.4"
             )
         }
 
@@ -167,5 +201,15 @@ class Tests {
 
         result.includes.filter { it.protocol == "file" }.map { File(it.toURI()).name } shouldBe List(4) { "include_${it + 1}.kt" }
         result.includes.filter { it.protocol != "file" }.size shouldBe 1
+    }
+
+    @Test
+    fun `test include detection - should not include dependency twice`() {
+      val result = resolveIncludes(File("test/resources/includes/dup_include/dup_include.kts"))
+
+      result.includes.map { File(it.toURI()).name } shouldBe listOf(
+            "dup_include_1.kt",
+            "dup_include_2.kt"
+        )
     }
 }